did another efactor i am so sorry
refactored to match the dotnet guidelines for sub projects
